requirements.txt
annotated-types==0.7.0
anyio==4.8.0
apeye==1.4.1
apeye-core==1.1.5
asttokens==2.0.8
backcall==0.2.0
bio==1.7.1
biopython==1.85
biothings_client==0.4.1
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
conllu==6.0.0
cymem==2.0.11
decorator==5.1.1
distro==1.9.0
domdf-python-tools==3.9.0
en_core_sci_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_md-0.5.4.tar.gz#sha256=7c8fc52542dd1452ffce00b045c1298e2c185b7cf84793f8e0ec941987c09808
entrypoints==0.4
exceptiongroup==1.2.2
executing==1.1.1
filelock==3.18.0
fsspec==2025.3.0
gprofiler-official==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.45
langchain-openai==0.3.8
langcodes==3.5.0
langsmith==0.3.13
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
mygene==3.2.2
natsort==8.4.0
networkx==3.4.2
nmslib @ file:///D:/bld/nmslib_1725861804688/work/python_bindings
numpy==1.26.4
openai==1.61.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.3
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pybind11 @ file:///D:/bld/pybind11-split_1730237329882/work
pybind11_global @ file:///D:/bld/pybind11-split_1730237329882/work
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.13.0
pypi-json==0.4.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==304
PyYAML==6.0.2
pyzmq==24.0.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rouge-metric==1.0.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
scispacy==0.5.5
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.5.1
sympy==1.13.1
tenacity==9.0.0
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.2
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.5
weasel==0.4.1
wrapt==1.17.2
zstandard==0.23.0
